refactor : date형식 변환, 종료일 계산 로직 파일 분리
